+++ src/HttpCache/ETagValueNormalizer/NoopETagValueNormalizer.php
@@ -7,7 +7,7 @@ namespace Yiisoft\HttpMiddleware\HttpCache\ETagValueNormalizer;
 /**
  * Returns ETag values unmodified. It can be used when ETag normalization is not required.
  */
-final class NullETagValueNormalizer implements ETagValueNormalizerInterface
+final class NoopETagValueNormalizer implements ETagValueNormalizerInterface
 {
     public function normalize(string $value): string
     {