+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Yii HTTP Middleware Change Log
 
-## 1.2.2 under development
+## 1.3.0 August 24, 2026
 
+- New #11: Add ETag value normalization in `HttpCacheMiddleware` via `ETagValueNormalizerInterface` with
+  `NoopETagValueNormalizer` and `SuffixETagValueNormalizer` implementations (@KalimeroMK)
 - New #29: Add `$keepHeadersOnStatusCode` and `$removedHeaders` constructor parameters to `RemoveBodyMiddleware` (@vjik)
 - New #29: Add `$removeOnStatusCode` constructor parameter to `ContentLengthMiddleware` (@vjik)
 - Bug #29: Remove `Content-Length` and `Transfer-Encoding` headers in `RemoveBodyMiddleware` when the body is
@@ -10,8 +12,6 @@
   not carry one (@vjik)
 - Bug #29: Add missing `103 Early Hints` status code to default status code lists in `RemoveBodyMiddleware` and
   `ContentLengthMiddleware` (@vjik)
-- New #11: Add ETag value normalization in `HttpCacheMiddleware` via `ETagValueNormalizerInterface` with
-  `NoopETagValueNormalizer` and `SuffixETagValueNormalizer` implementations (@KalimeroMK)
 
 ## 1.2.1 August 10, 2026